iOS autoPaginate ready for primetime
This builds upon what @realaboo did, primarily solving the bug he pointed out where iOS fires multiple times on first load. Basically, I just fire a flag indicating whether the first load is complete and disable infinite scrolling until that point, and voila! 

Also, there is some minor fixes in `_renderPaginationView()`. First wasn't even going to show since it relied on `pagination === true` and not also the `autoPaginate' prop. In also guarantee standard `pagination` views aren't shown when `autoPaginate` is set.  I supposed if you toggled both on, you might still have some issues, but it's not a big deal. It should be pretty obvious when those "issues" arise to toggle the one you're not using off.

Lastly, I'd commit this into @realaboo's branch, but he did his work before various changes in the latest repo. So credit goes to him, but also know that this is already combined with the latest code.  It should be good to go! Give it a whirl.