declarative api via `componentWillReceive` props + `rows` prop
this is primarily for usage with redux. the idea being that your `onFetch` prop will dispatch a redux action that ultimately will reduce the returns rows to the `rows` prop passed to GiftedListView, and when new rows are received, `componentWillReceiveProps` will serve the same purpose as the callback originally passed to `onFetch`. Therefore, onFetch doesn't ever need to call this callback (even though it will still receive it). When all rows are loaded just set the `allLoaded` prop to `true`. 

The only thing worth noting in the implementation is how the `beforeOptions` are kept in tack. Before they were passed to onFetch which ultimately passes them back, and possibly adds `allLoaded`, but now we store them on the instance at `this.beforeOptions`, resulting in calling code not having to worry about it at all, unless it wants to set `allLoaded` to true as described above. 

In addition, there's a `fetchOptions` prop which can be used to trigger a refresh, and then the calling of `onFetch` with the new params, which you can use to dispatch and make async queries which will ultimately reduce a new `rows` prop. The idea being that the interface used in its simplest most effective form consists of you just changing your `fetchOptions` :).