Tap to retry
Give option to user to retry a specific request.

Example Scenario:
-User is in 5th page.
-Network Request fails

Current Behaviour: This is not handled. User has to navigate to top and
pull to refresh the page.

Modified Behaviour: ‘Tap to retry’ view is shown which on tapping will
try to fetch the same request with same page no.